+++ src/main/java/org/betonquest/betonquest/quest/event/random/PickRandomEvent.java
@@ -4,21 +4,21 @@ import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.QuestTypeAPI;
 import org.betonquest.betonquest.api.quest.event.nullable.NullableEvent;
+import org.betonquest.betonquest.instruction.variable.VariableList;
 import org.betonquest.betonquest.instruction.variable.VariableNumber;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Iterator;
-import java.util.LinkedList;
 import java.util.List;
 
 /**
- * Fires randomly events by chance and amount.
+ * Fires randomly events by weight and amount.
  */
 public class PickRandomEvent implements NullableEvent {
     /**
-     * The events with there chance.
+     * The events with there weight.
      */
-    private final List<RandomEvent> events;
+    private final VariableList<RandomEvent> events;
 
     /**
      * The amount of events to fire.
@@ -34,11 +34,11 @@ public class PickRandomEvent implements NullableEvent {
     /**
      * Creates a new PickRandomEvent.
      *
-     * @param events       the events with there chance
+     * @param events       the events with there weight
      * @param amount       the amount of events to fire
      * @param questTypeAPI the Quest Type API
      */
-    public PickRandomEvent(final List<RandomEvent> events, @Nullable final VariableNumber amount, final QuestTypeAPI questTypeAPI) {
+    public PickRandomEvent(final VariableList<RandomEvent> events, @Nullable final VariableNumber amount, final QuestTypeAPI questTypeAPI) {
         this.events = events;
         this.amount = amount;
         this.questTypeAPI = questTypeAPI;
@@ -46,24 +46,21 @@ public class PickRandomEvent implements NullableEvent {
 
     @Override
     public void execute(@Nullable final Profile profile) throws QuestException {
-        final List<ResolvedRandomEvent> resolvedEvents = new LinkedList<>();
-        for (final RandomEvent randomEvent : events) {
-            resolvedEvents.add(randomEvent.resolveFor(profile));
-        }
-        double total = resolvedEvents.stream().mapToDouble(ResolvedRandomEvent::chance).sum();
+        final List<RandomEvent> resolvedEvents = events.getValue(profile);
+        double total = resolvedEvents.stream().mapToDouble(RandomEvent::weight).sum();
 
         int pick = this.amount == null ? 1 : this.amount.getValue(profile).intValue();
         while (pick > 0 && !resolvedEvents.isEmpty()) {
             pick--;
             double random = Math.random() * total;
-            final Iterator<ResolvedRandomEvent> iterator = resolvedEvents.iterator();
+            final Iterator<RandomEvent> iterator = resolvedEvents.iterator();
             while (iterator.hasNext()) {
-                final ResolvedRandomEvent event = iterator.next();
-                random -= event.chance();
+                final RandomEvent event = iterator.next();
+                random -= event.weight();
                 if (random < 0) {
                     questTypeAPI.event(profile, event.eventID());
                     iterator.remove();
-                    total -= event.chance();
+                    total -= event.weight();
                     break;
                 }
             }